10th March 2021 - Push to Main
I am merging the code here with the main branch to bring in the code together as one block as there are a couple of scripts that are now distributed across several branches.

Here I am uploading new scripts for the selection of the 6 points used for ODMR scans.

I also have a couple of lock-in scripts as well.