7th July 2021
I am uploading the currently used scripts to the online github. The reason for this upload is to include the modifications that have had to be made as a result of the labview triggering the first frequency scan point before the first point is recorded. This is a hangover from previous methods of doing the experiments and should be easily fixed.

The way I deal with this is to use the np.roll() function in python, which brings the last item in the data set to the first position and moves all other data points back one position in the array.